requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.11.0
attrs==24.2.0
biopython==1.85
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.18
dotenv==0.9.9
fastapi==0.114.0
frozenlist==1.4.1
future==1.0.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.8
Levenshtein==0.27.1
line-bot-sdk==3.18.0
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.0.5
pip==25.1
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
requests==2.32.3
setuptools==78.1.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.38.5
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.2.2
uvicorn==0.30.6
wheel==0.45.1
wrapt==1.16.0
yarl==1.20.1
gunicorn==23.00